+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+### Added
+
+ - Statistics: measure the total number of allocated & freed nodes.
+
 ## 0.6.0 - 2025-07-25
 
 ### Changed